requirements.txt
asgiref==3.4.1
Django==3.2.5
django-etc==1.3.0
django-hitcount==1.3.3
django-taggit==1.5.1
Pillow==8.3.1
pytz==2021.1
sqlparse==0.4.1
